+++ lib/xcodeproj.rb
@@ -20,13 +20,6 @@ module Xcodeproj
   autoload :Workspace,        'xcodeproj/workspace'
   autoload :XCScheme,         'xcodeproj/scheme'
   autoload :XcodebuildHelper, 'xcodeproj/xcodebuild_helper'
-
-  # TODO: Delete me (compatibility with CocoaPods 0.33.1)
-  def self.read_plist(path)
-    PlistHelper.read(path)
-  end
-
-  def self.write_plist(hash, path)
-    PlistHelper.write(hash, path)
-  end
 end
+
+require 'xcodeproj/ext'