+++ lib/xcodeproj/project.rb
@@ -172,7 +172,7 @@ module Xcodeproj
     #
     def initialize_from_file
       pbxproj_path = path + 'project.pbxproj'
-      plist = Xcodeproj::PlistHelper.read(pbxproj_path.to_s)
+      plist = Xcodeproj.read_plist(pbxproj_path.to_s)
       root_object_uuid = plist['rootObject']
       root_object.remove_referrer(self) if root_object
       @root_object = new_from_plist(root_object_uuid, plist['objects'], self)
@@ -302,7 +302,7 @@ module Xcodeproj
       save_path ||= path
       FileUtils.mkdir_p(save_path)
       file = File.join(save_path, 'project.pbxproj')
-      Xcodeproj::PlistHelper.write(to_hash, file)
+      Xcodeproj.write_plist(to_hash, file)
       fix_encoding(file)
       XCProjHelper.touch(save_path) unless disable_xcproj?
     end
@@ -701,7 +701,7 @@ module Xcodeproj
       end
 
       xcschememanagement_path = schemes_dir + 'xcschememanagement.plist'
-      Xcodeproj::PlistHelper.write(xcschememanagement, xcschememanagement_path)
+      Xcodeproj.write_plist(xcschememanagement, xcschememanagement_path)
     end
 
     #-------------------------------------------------------------------------#